Improved cache resize policy: eXist uses dynamic caches to adopt to different loads. It is thus important to determine if a page should be grown or shrinked in order to increase cache efficiency. So far, a cache was resized based on the number of page replacements. However, this is a bad measure as replacements can be positive as well as negative.
The new implementation bases the resize decision on the thrashing factor of a cache. Thrashing occurs if a page is replaced from the cache and has to be reloaded shortly after. For the B+-tree pages, we normally don't want any trashing. For the data pages, a certain amount of thrashing is ok.

Additionally, a huge cache can now be shrinked dynamically if a smaller cache requests a resize. Some caches (words.dbx) tend to grow very fast and consume the whole cache space, so there will be no space left for the slower-growing caches. This may lead to a bottleneck if the amount of stored data increases.

svn path=/trunk/eXist-1.0/; revision=1637

+++ src/org/exist/storage/btree/BTree.java
@@ -147,7 +147,7 @@ public class BTree extends Paged {
     /** Cache of BTreeNode(s) */
     protected Cache cache;
     
-    protected int growthThreshold;
+    protected double growthThreshold;
     
     /** Size of BTreeNode cache */
     protected int buffers;
@@ -162,7 +162,7 @@ public class BTree extends Paged {
     
     protected boolean isTransactional;
     
-	protected BTree(BrokerPool pool, byte fileId, boolean transactional, CacheManager cacheManager, int growthThreshold) {
+	protected BTree(BrokerPool pool, byte fileId, boolean transactional, CacheManager cacheManager, double growthThreshold) {
 		super();
         this.cacheManager = cacheManager;
         this.buffers = cacheManager.getDefaultInitialSize();
@@ -176,7 +176,7 @@ public class BTree extends Paged {
             logManager = pool.getTransactionManager().getJournal();
 	}
 
-	public BTree(BrokerPool pool, byte fileId, boolean transactional, CacheManager cacheManager, File file, int growthThreshold) {
+	public BTree(BrokerPool pool, byte fileId, boolean transactional, CacheManager cacheManager, File file, double growthThreshold) {
         this(pool, fileId, transactional, cacheManager, growthThreshold);
 		setFile(file);
 	}
